% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexmisc.R
\docType{package}
\name{alexmisc}
\alias{alexmisc}
\title{alexmisc: Alexander Brenning's Miscellaneous Functions}
\description{
This package is a collection of miscellaneous utility
functions, e.g. functions to make code more robust (and
the use of try-catch functions more convenient),
utility functions for reading web pages (including
Wikipedia), and some other fun stuff.
}
