% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_images.R
\name{download_new_wiki_image}
\alias{download_new_wiki_image}
\title{Download Wikipedia image if it doesn't already exist}
\usage{
download_new_wiki_image(
  name,
  destfile = gsub(" ", "_", name),
  destfolder = ".",
  ...
)
}
\arguments{
\item{name}{Name of Wikipedia entry.}

\item{destfile}{Destination folder and file name for saving image. If \code{NULL} (default), use name of Wikipedia page.}

\item{destfolder}{Destination folder and file name for saving image. If \code{NULL} (default), use name of Wikipedia page.}

\item{...}{additional arguments passed to \code{\link[=download_wiki_image]{download_wiki_image()}}.}
}
\value{
A list containing image metadata, as returned by \code{\link[=download_wiki_image]{download_wiki_image()}}.
}
\description{
Download Wikipedia image if it doesn't already exist
}
