% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_images.R
\name{download_wiki_image}
\alias{download_wiki_image}
\title{Download image from Wikipedia}
\usage{
download_wiki_image(
  name,
  destfolder = ".",
  destfile = NULL,
  wiki = c("en", "de")[1],
  base_url = NULL,
  filter = NULL,
  which = c("first", "last", "random")[1],
  save_metadata = TRUE,
  verbose = FALSE,
  quiet_download = !verbose
)
}
\arguments{
\item{name}{Name of Wikipedia entry.}

\item{destfolder, destfile}{Destination folder and file name for saving image. If \code{NULL} (default), use name of Wikipedia page.}

\item{wiki}{Which Wikipedia to use: English (\code{"en"}, default), or German (\code{"de"}).}

\item{base_url}{Wikipedia base URL, which depends on \code{wiki}.}

\item{filter}{Defaults to \code{NULL}, which will pick the appropriate filter. E.g., \code{"/wiki/File:"}, depending on language.}

\item{which}{Determines which picture to choose: Either \code{"first"} (default), \code{"last"}, \code{"random"}, or a numeric value indicating the picture number.}

\item{save_metadata}{If \code{TRUE} (default and recommended), save extracted metadata (e.g., source URL, author, licence, if possible) to a \code{.rds} file with same file name as image file. The metadata is stored in a list format.}

\item{verbose}{Logical; controls verbosity.}

\item{quiet_download}{Passed to \code{\link[=download.file]{download.file()}} as its \code{quiet} argument. Defaults to \code{!verbose}.}
}
\value{
Invisible returns a list with metadata such as source URL, author, licence, if possible.
}
\description{
Download image from Wikipedia
}
\examples{
# res <- download_wiki_image("ArcGIS",
#                            destfile = "test", destfolder = "figures")
# res <- download_wiki_image("Stoppschild", wiki = "de",
#                            destfile = "test", destfolder = "figures")
# res <- download_wiki_image("Lake Winnipesaukee",
#                            destfile = "test", destfolder = "figures")
# res <- download_wiki_image("E10 (Kraftstoff)", wiki = "de", which = 2,
#                            destfolder = "figures")
}
