% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{failed}
\alias{failed}
\title{Determine if \code{try()} call failed}
\usage{
failed(x)
}
\arguments{
\item{x}{Result of a \code{\link[=try]{try()}} call.}
}
\value{
\code{TRUE} if \code{x} is of class \code{"try-error"}, indicating
an error.
}
\description{
Determine if \code{try()} call failed
}
\examples{
failed(try(sqrt("a")))
failed(try(sqrt(2)))
}
\seealso{
\code{\link[=try]{try()}}
}
