% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_wiki.R
\name{geocode_wiki_town}
\alias{geocode_wiki_town}
\alias{geocode_wiki_towns}
\title{Geocode town name using (German) Wikipedia}
\usage{
geocode_wiki_town(name, wiki = "de", digits = 5, delay = 0.1)

geocode_wiki_towns(names, wiki = "de", digits = 5, delay = 1)
}
\arguments{
\item{name, names}{Name of town as it appears in its Wikipedia URL, e.g. \code{"Halle_(Saale)"} or .}

\item{wiki}{Identifies Wikipedia language edition, e.g. \code{"de"} for German (default) or \code{"en"} for English.}

\item{digits}{Round lat/lon to this many decimal places.}

\item{delay}{Delay (in seconds) after Wikipedia call.}
}
\value{
A list containing the \code{name}, \code{latitude} and \code{longitude} of the requested town.
Latitude and longitude are \code{NA} if the attempt failed.
}
\description{
A quick and dirty attempt at geocoding places using Wikipedia.
}
\section{Functions}{
\itemize{
\item \code{geocode_wiki_towns}: Iterate \code{geocode_wiki_town} over multiple towns
}}

\examples{
geocode_wiki_town("Jena")
geocode_wiki_town("Halle_(Saale)") # with underscore
geocode_wiki_town("Nürnberg") # with umlaut
}
