% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{if_missing}
\alias{if_missing}
\title{Replace NULL and NA values}
\usage{
if_missing(x, y)
}
\arguments{
\item{x}{an object, usually an atomic one}

\item{y}{value to be returned if \code{x} is \code{NULL}, of length 0,
or if its first element is \code{NA}}
}
\value{
\code{y}, if \code{x} is \code{NULL}, length 0, or if \code{x[1]} is \code{NA}; \code{x} otherwise.
}
\description{
Replace NULL and NA values
}
\examples{
if_missing(NULL, "n/a")
if_missing(NA, "n/a")
if_missing(c(), "n/a")
if_missing("something", "n/a")
}
