% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_images.R
\name{image_caption}
\alias{image_caption}
\title{Simple image caption from metadata}
\usage{
image_caption(
  x,
  use_name = TRUE,
  use_date = FALSE,
  nchar_description = 30,
  language = c("de", "en")[1]
)
}
\arguments{
\item{x}{Image metadata as returned by \code{\link[=download_wiki_image]{download_wiki_image()}}. The \code{description} field will be used to create a figure caption.}

\item{use_name}{Include the author's name in figure caption? (Default: \code{TRUE}).}

\item{use_date}{Include the image date in figure caption? (Default: \code{FALSE}).}

\item{nchar_description}{Maximum number of characters for the description.}

\item{language}{Either \code{"de"} (German caption; the current default) or \code{"en"} (English version).}
}
\value{
A single character string with an image description and basic provenance information.
}
\description{
Simple image caption from metadata
}
