% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_images.R
\name{include_md_image}
\alias{include_md_image}
\title{Include an image file in an RMarkdown document}
\usage{
include_md_image(
  file,
  path = NULL,
  caption = NULL,
  default_caption = "",
  width = NULL,
  include_suppl = TRUE,
  fontsize = -2
)
}
\arguments{
\item{file, path}{File name and folder of image to be included in the Markdown document. It is recommended to use \code{\link[here:here]{here::here()}} to specify the path.}

\item{caption}{Figure caption. If \code{"NULL"} and metadata exists, use \code{\link[=image_caption]{image_caption()}} to construct an image caption.}

\item{default_caption}{Default caption, if \code{caption} argument is \code{NULL}.}

\item{width}{Image width, e.g. \code{"5cm"}. If \code{NULL} (default), don't specify width in Markdown.}

\item{include_suppl}{Include image file? Defaults to \code{TRUE}.}

\item{fontsize}{HTML font size, relative to current standard setting. Default: \code{-2}.}
}
\value{
Markdown code to be included in the document. The function is also called for its side effect of calling \code{\link[exams:include_supplement]{exams::include_supplement()}}.
}
\description{
Include an image file in an RMarkdown document
}
