% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_null_na}
\alias{is_null_na}
\title{Check if object is \code{NULL} or all \code{NA}s}
\usage{
is_null_na(x)
}
\arguments{
\item{x}{an atomic object or vector}
}
\value{
\code{TRUE}, if \code{x} is either \code{NULL} or \code{NA}; \code{FALSE} otherwise.
}
\description{
Check if object is \code{NULL} or all \code{NA}s
}
\examples{
is_null_na(NULL)
is_null_na(NA)
is_null_na(c(NA, NA, NA))
is_null_na(c("something", NA))
}
