% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_images.R
\name{my_abbreviate}
\alias{my_abbreviate}
\title{Abbreviation by ellipsis}
\usage{
my_abbreviate(x, nc, replacement = "...")
}
\arguments{
\item{x}{Character string to be abbreviated.}

\item{nc}{Maximum number of characters of the result.}

\item{replacement}{Replacement, defaults to \code{"..."}.}
}
\value{
An abbreviated character string of maximum \code{nc} characters length.
}
\description{
Abbreviation by ellipsis
}
\examples{
my_abbreviate("Hello, World!", nc = 8)
}
