% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykrige.R
\name{predict.myidw}
\alias{predict.myidw}
\title{Predict method for \code{myidw} objects.}
\usage{
\method{predict}{myidw}(object, newdata, locations = ~x + y, nmax = object$nmax, ...)
}
\arguments{
\item{object, newdata, locations, nmax}{See \code{\link[gstat:krige]{gstat::idw()}}.}

\item{...}{See \code{\link[gstat:krige]{gstat::idw()}}.}
}
\value{
The predict method returns only the predicted values as a numeric vector.
}
\description{
Predict method for \code{myidw} objects.
}
