% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykrige.R
\name{predict.mykrige}
\alias{predict.mykrige}
\title{Predict method for \code{mykrige} objects}
\usage{
\method{predict}{mykrige}(object, newdata, nmax = object$nmax, ...)
}
\arguments{
\item{object, newdata}{See \code{\link[gstat:krige]{gstat::krige()}}.}

\item{nmax}{number of nearest observations to be used in local kriging; defaults to value specified when creating the \code{mykrige} object.}

\item{...}{See \code{\link[gstat:krige]{gstat::idw()}}}
}
\value{
The predict method return only the predicted values as a numeric vector.
}
\description{
Predict method for \code{mykrige} objects
}
