% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recursive_file_exists}
\alias{recursive_file_exists}
\title{Recursive version of \code{file.exists()}}
\usage{
recursive_file_exists(x, path = ".")
}
\arguments{
\item{x}{File name}

\item{path}{Folder in which to start the recursive search.}
}
\value{
\code{TRUE} if at least one file matching \code{x} was
found, and \code{FALSE} otherwise.
}
\description{
Recursive version of \code{file.exists()}
}
\details{
Currently does not support regular expressions or
wildcards; dot (for file extensions etc.) are dealt with.
}
\examples{
# recursive_file_exists("I_dont_exist.R")
}
