% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_null}
\alias{replace_null}
\title{Replace \code{NULL} values with \code{NA}}
\usage{
replace_null(x)
}
\arguments{
\item{x}{An object, or \code{NULL}}
}
\value{
\code{NA} if \code{x} is \code{NULL}, \code{x} otherwise.
}
\description{
Replace \code{NULL} values with \code{NA}
}
