% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_umlaute}
\alias{replace_umlaute}
\title{Replace German umlaute with \code{"ae"} and the like}
\usage{
replace_umlaute(x)
}
\arguments{
\item{x}{Character string containing umlaute}
}
\value{
Character string with umlaute replaced
}
\description{
Replace German umlaute with \code{"ae"} and the like
}
\examples{
replace_umlaute("Jenalöbnitz")
replace_umlaute("Jenaprießnitz")
}
