% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykrige.R
\name{svgm_spherical}
\alias{svgm_spherical}
\title{Spherical semivariogram}
\usage{
svgm_spherical(x, range, psill, nugget = 0)
}
\arguments{
\item{x}{Lag distance (numeric vector)}

\item{range}{Range}

\item{psill}{Partial sill (i.e. excluding nugget)}

\item{nugget}{Nugget semivariance}
}
\value{
Semivariances at the specified distances.
}
\description{
Spherical semivariogram
}
