% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{try_read_html}
\alias{try_read_html}
\title{Try and retry reading a webpage}
\usage{
try_read_html(
  expose_url,
  delay_before = 0,
  delay_after = 1,
  max_tries = 2,
  interval = max(delay_before, delay_after),
  ...
)
}
\arguments{
\item{expose_url}{URL}

\item{delay_before, delay_after}{Delay (in seconds) before and/or after trying to read the webpage}

\item{max_tries}{Maximum number of attempts}

\item{interval}{Interval between the attempts; defaults to the maximum of the two delay argument.}

\item{...}{Additional arguments for \code{\link[xml2:read_xml]{xml2::read_html()}}}
}
\value{
The result of \code{\link[xml2:read_xml]{xml2::read_html()}}, or an object of class \code{"try-error"} in case of error.
}
\description{
Try reading a webpage possibly multiple times in case of error, with delays before, between, and after the attempts.
}
