% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{try_retry}
\alias{try_retry}
\title{Convenience function that tr}
\usage{
try_retry(expr, when = "error", ...)
}
\arguments{
\item{expr}{An expression to be evaluated}

\item{when}{Condition under which to retry, see \code{\link[retry:retry]{retry::retry()}}}

\item{...}{Additional arguments for \code{\link[retry:retry]{retry::retry()}}}
}
\value{
The result of the \code{\link[=try]{try()}} call, i.e. the value of
the expression being evaluated, or an object of class
\code{"try-error"}.
}
\description{
Convenience function that tr
}
\examples{
failed(try_retry(sqrt("a")))
# This one won't get any better just by repeating its
# evaluation but maybe if a webpage can't be read,
# it's worth a second try...
}
\seealso{
\code{\link[=try_read_html]{try_read_html()}}
}
