% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykrige.R
\name{myidw}
\alias{myidw}
\alias{predict.myidw}
\alias{predict.mykrige}
\title{IDW interpolation wrapper}
\usage{
myidw(formula, data)

\method{predict}{myidw}(object, newdata, locations = ~x + y, nmax = 7, ...)

\method{predict}{mykrige}(object, newdata, nmax = 7, ...)
}
\arguments{
\item{formula, data, locations}{See \code{\link[gstat:krige]{gstat::idw()}}.}
}
\value{
\code{myidw} returns a list with the formula and data, i.e. an object of class \code{myidw}. The predict method return only the predicted values as a numeric vector.
}
\description{
IDW interpolation wrapper
}
\section{Methods (by generic)}{
\itemize{
\item \code{predict}: Predict method for \code{myidw} objects.

\item \code{predict}: Predict method for \code{mykrige} objects.
}}

