% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykrige.R
\name{mykrige}
\alias{mykrige}
\title{Kriging interpolation wrapper}
\usage{
mykrige(
  formula,
  data,
  locations = ~x + y,
  range = 500,
  nsratio = 0.25,
  fixed = FALSE
)
}
\arguments{
\item{formula, data, locations}{See \code{\link[gstat:krige]{gstat::krige()}}.}

\item{range}{Range parameter of spherical semivariogram}

\item{nsratio}{Nugget-to-sill ratio of spherical semivariogram}

\item{fixed}{Determines if the semivariogram parameters should be fixed or not; default: \code{FALSE}.}
}
\value{
\code{mykrige} returns a list with the formula and data, i.e. an object of class \code{mykrige}.
}
\description{
Kriging interpolation wrapper
}
