% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mykrige.R
\name{predict.mykrige}
\alias{predict.mykrige}
\title{Predict method for \code{mykrige} objects}
\usage{
\method{predict}{mykrige}(object, newdata, nmax = 30, ...)
}
\arguments{
\item{object, newdata, nmax, ...}{See \code{\link[gstat:krige]{gstat::krige()}}.}
}
\value{
The predict method return only the predicted values as a numeric vector.
}
\description{
Predict method for \code{mykrige} objects
}
